requirements.txt
bcrypt
pycryptodome
requests
getmac
